V2

+++ core/datastore/src/main/java/com/jparkbro/datastore/RecentAnimeDataStore.kt
@@ -1,8 +1,8 @@
 package com.jparkbro.datastore
 
 interface RecentAnimeDataStore {
-    suspend fun saveRecentAnime(animeId: Int): Result<Unit>
-    suspend fun loadRecentAnime(): Result<Int>
+    suspend fun saveRecentAnime(animeId: Long): Result<Unit>
+    suspend fun loadRecentAnime(): Result<Long>
 
     suspend fun clearRecentAnime(): Result<Unit>
 }
